requirements.txt
adagio==0.2.6
amqp==5.3.1
appdirs==1.4.4
asgiref==3.9.2
billiard==4.2.2
celery==5.5.3
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.3.0
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cloudpickle==3.1.1
Django==5.2.6
django-redis==6.0.0
flatbuffers==25.2.10
fs==2.4.16
fsspec==2025.9.0
fugue==0.9.1
h11==0.16.0
h5py==3.14.0
idna==3.10
jh2==5.0.10
kombu==5.5.4
llvmlite==0.45.1
narwhals==2.6.0
niquests==3.15.2
numba==0.62.1
numpy==2.3.3
openmeteo_requests==1.7.3
openmeteo_sdk==1.21.2
packaging==25.0
pandas==2.3.3
plotly==6.3.1
prompt_toolkit==3.0.52
psycopg2==2.9.10
pvlib==0.13.1
pyarrow==21.0.0
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2025.2
qh3==1.5.5
redis==6.4.0
requests==2.32.5
scipy==1.16.2
setuptools==80.9.0
six==1.17.0
sqlparse==0.5.3
tqdm==4.67.1
triad==0.9.8
tzdata==2025.2
urllib3==2.5.0
urllib3-future==2.14.902
vine==5.1.0
wassima==2.0.2
wcwidth==0.2.14
